--- ProjectIdle/app/src/com/idel/gui/Application.java
@@ -1,20 +1,0 @@
-package com.idel.gui;
-
-import java.awt.Font;
-
-public class Application {
-
-	public static Font labelFont = new Font("DungGeunMo", Font.PLAIN, 25);
-	public static Font middleFont = new Font("DungGeunMo", Font.PLAIN, 50);
-	public static Font titleFont = new Font("DungGeunMo", Font.BOLD, 90);
-	public static Font littlebigFont = new Font("DungGeunMo", Font.BOLD, 55);
-
-	public static void main(String[] args) {
-
-//		new GameMain(); 
-		new FindPasswordPage();
-		// 허희만
-
-	}
-
-}
