+++ ProjectIdle/app/src/user/player/info/UserInfoPage.java
@@ -1,4 +1,4 @@
-package com.idel.gui;
+package user.player.info;
 
 import java.awt.Color;
 import java.awt.Font;
@@ -11,6 +11,8 @@ import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JTextField;
+
+import user.player.signup.LoginPage;
 
 public class UserInfoPage extends JFrame {
 	private JFrame myPage;
