+++ ProjectIdle/app/src/user/player/realplay/PlayMain.java
@@ -1,4 +1,4 @@
-package com.idel.gui;
+package user.player.realplay;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
@@ -8,6 +8,8 @@ import javax.swing.JButton;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
+
+import user.player.gamemain.GameMain;
 
 public class PlayMain extends JFrame{
