+++ ProjectIdle/app/src/user/player/store/receip/RecipeStoragePage.java
@@ -1,4 +1,4 @@
-package com.idel.gui;
+package user.player.store.receip;
 
 import java.awt.Font;
 import java.awt.Image;
@@ -10,6 +10,8 @@ import javax.swing.JButton;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
+
+import user.player.storage.StoragePage;
 
 public class RecipeStoragePage extends JFrame {
